feat: add GetBestStudents function

+++ model/Mark.go
@@ -3,6 +3,6 @@ package model
 type Mark struct {
 	ID        uint64 `json:"id" sql:"AUTO_INCREMENT" gorm:"primary_key"`
 	Value     float64
-	StudentID uint
-	ClassID   uint
+	StudentID uint64
+	ClassID   uint64
 }
